+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Test data contract against dataframes / temporary views (#175)
 
+### Fixed
+- AVRO export: Logical Types should be nested (#233)
+
 ## [0.10.6] - 2024-05-29
 
 ### Fixed